requirements.txt
diffusers>=0.25.0
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.25.0
matplotlib>=3.7.0
Pillow>=10.0.0
boto3>=1.26.0
python-dotenv>=0.19.0
openai>=1.0.0
